Break circular dependencies for query object
